moved test to a joing tes_path.py module under test_kinematics
