+++ tests/test_unit/test_kinematics/test_path.py
@@ -1,10 +1,168 @@
-"""Tests for compute_path_straightness."""
+from contextlib import nullcontext as does_not_raise
 
 import numpy as np
 import pytest
 import xarray as xr
 
-from movement.kinematics.path import compute_path_straightness
+from movement.kinematics import compute_path_length, compute_path_straightness
+
+time_points_value_error = pytest.raises(
+    ValueError,
+    match="At least 2 time points are required to compute path length",
+)
+
+
+@pytest.mark.parametrize(
+    "start, stop, expected_exception",
+    [
+        # full time ranges
+        (None, None, does_not_raise()),
+        (0, None, does_not_raise()),
+        (0, 9, does_not_raise()),
+        (0, 10, does_not_raise()),  # xarray.sel will truncate to 0, 9
+        (-1, 9, does_not_raise()),  # xarray.sel will truncate to 0, 9
+        # partial time ranges
+        (1, 8, does_not_raise()),
+        (1.5, 8.5, does_not_raise()),
+        (2, None, does_not_raise()),
+        # Empty time ranges
+        (9, 0, time_points_value_error),  # start > stop
+        ("text", 9, time_points_value_error),  # invalid start type
+        # Time range too short
+        (0, 0.5, time_points_value_error),
+    ],
+)
+def test_path_length_across_time_ranges(
+    valid_poses_dataset, start, stop, expected_exception
+):
+    """Test path length computation for a uniform linear motion case,
+    across different time ranges.
+
+    The test dataset ``valid_poses_dataset``
+    contains 2 individuals ("id_0" and "id_1"), moving
+    along x=y and x=-y lines, respectively, at a constant velocity.
+    At each frame they cover a distance of sqrt(2) in x-y space, so in total
+    we expect a path length of sqrt(2) * num_segments, where num_segments is
+    the number of selected frames minus 1.
+    """
+    position = valid_poses_dataset.position
+    with expected_exception:
+        path_length = compute_path_length(position, start=start, stop=stop)
+        assert path_length.name == "path_length"
+        assert path_length.long_name == "Path Length"
+
+        # Expected number of segments (displacements) in selected time range
+        num_segments = 9  # full time range: 10 frames - 1
+        start = max(0, start) if start is not None else 0
+        stop = min(9, stop) if stop is not None else 9
+        if start is not None:
+            num_segments -= np.ceil(max(0, start))
+        if stop is not None:
+            stop = min(9, stop)
+            num_segments -= 9 - np.floor(min(9, stop))
+
+        expected_path_length = xr.DataArray(
+            np.ones((3, 2)) * np.sqrt(2) * num_segments,
+            dims=["keypoints", "individuals"],
+            coords={
+                "keypoints": position.coords["keypoints"],
+                "individuals": position.coords["individuals"],
+            },
+        )
+        xr.testing.assert_allclose(path_length, expected_path_length)
+
+
+@pytest.mark.parametrize(
+    "nan_policy, expected_path_lengths_id_0, expected_exception",
+    [
+        (
+            "ffill",
+            np.array([np.sqrt(2) * 9, np.sqrt(2) * 8, np.nan]),
+            does_not_raise(),
+        ),
+        (
+            "scale",
+            np.array([np.sqrt(2) * 9, np.sqrt(2) * 9, np.nan]),
+            does_not_raise(),
+        ),
+        (
+            "invalid",  # invalid value for nan_policy
+            np.zeros(3),
+            pytest.raises(ValueError, match="Invalid value for nan_policy"),
+        ),
+    ],
+)
+def test_path_length_with_nan(
+    valid_poses_dataset_with_nan,
+    nan_policy,
+    expected_path_lengths_id_0,
+    expected_exception,
+):
+    """Test path length computation for a uniform linear motion case,
+    with varying number of missing values per individual and keypoint.
+    Because the underlying motion is uniform linear, the "scale" policy should
+    perfectly restore the path length for individual "id_0" to its true value.
+    The "ffill" policy should do likewise if frames are missing in the middle,
+    but will not "correct" for missing values at the edges.
+    """
+    position = valid_poses_dataset_with_nan.position
+    with (
+        pytest.warns(UserWarning, match="The result may be unreliable"),
+        expected_exception,
+    ):
+        path_length = compute_path_length(position, nan_policy=nan_policy)
+        assert path_length.name == "path_length"
+        # Get path_length for individual "id_0" as a numpy array
+        path_length_id_0 = path_length.sel(individuals="id_0").values
+        # Check them against the expected values
+        np.testing.assert_allclose(
+            path_length_id_0, expected_path_lengths_id_0
+        )
+
+
+# Regex patterns to match the warning messages
+exclude_id_1_and_left = r"(?s)(?!.*id_1)(?!.*left)"
+include_threshold_100 = r".*The result may be unreliable.*right.*id_0.*10/10.*"
+include_threshold_20 = (
+    r".*The result may be unreliable.*centroid.*right.*id_0.*3/10.*10/10.*"
+)
+
+
+@pytest.mark.parametrize(
+    "nan_warn_threshold, expected_exception",
+    [
+        (
+            1,
+            pytest.warns(
+                UserWarning,
+                match=f"{exclude_id_1_and_left}{include_threshold_100}",
+            ),
+        ),
+        (
+            0.2,
+            pytest.warns(
+                UserWarning,
+                match=f"{exclude_id_1_and_left}{include_threshold_20}",
+            ),
+        ),
+        (-1, pytest.raises(ValueError, match="between 0 and 1")),
+    ],
+)
+def test_path_length_nan_warn_threshold(
+    valid_poses_dataset_with_nan, nan_warn_threshold, expected_exception
+):
+    """Test that a warning is raised with matching message containing
+    information on the individuals and keypoints whose number of missing
+    values exceeds the given threshold or that an error is raised
+    when the threshold is invalid.
+    """
+    position = valid_poses_dataset_with_nan.position
+    with expected_exception:
+        result = compute_path_length(
+            position, nan_warn_threshold=nan_warn_threshold
+        )
+        assert result.name == "path_length"
+
 
 # ─────────────────────────────────────────────
 # Test dataset factories
@@ -161,7 +319,7 @@ def test_straightness_output_name_and_units():
     """Output should have correct name and units attributes."""
     result = compute_path_straightness(make_straight_line())
     assert result.name == "straightness_index"
-    assert result.attrs.get("units") == "dimensionless"
+    assert result.long_name == "Path Straightness Index"
 
 
 # ─────────────────────────────────────────────
